requirements.txt
groq
requests
python-dotenv
flask
flask-cors
beautifulsoup4
